format code

+++ src/main/java/com/olivia/peanut/aps/service/impl/ApsRollingForecastFactoryCapacityServiceImpl.java
@@ -1,5 +1,7 @@
 package com.olivia.peanut.aps.service.impl;
 
+import static com.olivia.sdk.utils.FieldUtils.getField;
+
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.util.ReflectUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
@@ -10,7 +12,13 @@ import com.github.yulichang.wrapper.MPJLambdaWrapper;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.collect.Maps;
-import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.*;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityDto;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityExportQueryPageListReq;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityInsertReq;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityInsertRes;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityQueryListReq;
+import com.olivia.peanut.aps.api.entity.apsRollingForecastFactoryCapacity.ApsRollingForecastFactoryCapacityQueryListRes;
 import com.olivia.peanut.aps.mapper.ApsRollingForecastFactoryCapacityMapper;
 import com.olivia.peanut.aps.model.ApsRollingForecastFactoryCapacity;
 import com.olivia.peanut.aps.service.ApsRollingForecastFactoryCapacityService;
@@ -19,22 +27,27 @@ import com.olivia.peanut.base.service.BaseTableHeaderService;
 import com.olivia.peanut.base.service.FactoryService;
 import com.olivia.peanut.util.SetNamePojoUtils;
 import com.olivia.sdk.service.SetNameService;
-import com.olivia.sdk.utils.*;
+import com.olivia.sdk.utils.$;
+import com.olivia.sdk.utils.DateUtils;
+import com.olivia.sdk.utils.DynamicsPage;
+import com.olivia.sdk.utils.FieldUtils;
+import com.olivia.sdk.utils.LambdaQueryUtil;
 import com.olivia.sdk.utils.model.YearMonth;
 import jakarta.annotation.Resource;
+import java.lang.reflect.Field;
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.aop.framework.AopContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.lang.reflect.Field;
-import java.time.LocalDate;
-import java.util.*;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-import static com.olivia.sdk.utils.FieldUtils.getField;
 
 /**
  * 滚动预测(ApsRollingForecastFactoryCapacity)表服务实现类
@@ -45,10 +58,12 @@ import static com.olivia.sdk.utils.FieldUtils.getField;
 @Slf4j
 @Service("apsRollingForecastFactoryCapacityService")
 @Transactional
-public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseServiceImpl<ApsRollingForecastFactoryCapacityMapper, ApsRollingForecastFactoryCapacity> implements
+public class ApsRollingForecastFactoryCapacityServiceImpl extends
+    MPJBaseServiceImpl<ApsRollingForecastFactoryCapacityMapper, ApsRollingForecastFactoryCapacity> implements
     ApsRollingForecastFactoryCapacityService {
 
-  final static Cache<String, Map<String, String>> cache = CacheBuilder.newBuilder().maximumSize(100).expireAfterWrite(30, TimeUnit.MINUTES).build();
+  final static Cache<String, Map<String, String>> cache = CacheBuilder.newBuilder().maximumSize(100)
+      .expireAfterWrite(30, TimeUnit.MINUTES).build();
 
   @Resource
   BaseTableHeaderService tableHeaderService;
@@ -59,17 +74,21 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
 
   // 以下为私有对象封装
 
-  public @Override ApsRollingForecastFactoryCapacityQueryListRes queryList(ApsRollingForecastFactoryCapacityQueryListReq req) {
+  public @Override ApsRollingForecastFactoryCapacityQueryListRes queryList(
+      ApsRollingForecastFactoryCapacityQueryListReq req) {
 
     MPJLambdaWrapper<ApsRollingForecastFactoryCapacity> q = getWrapper(req.getData());
     List<ApsRollingForecastFactoryCapacity> list = this.list(q);
 
-    List<ApsRollingForecastFactoryCapacityDto> dataList = list.stream().map(t -> $.copy(t, ApsRollingForecastFactoryCapacityDto.class)).collect(Collectors.toList());
+    List<ApsRollingForecastFactoryCapacityDto> dataList = list.stream()
+        .map(t -> $.copy(t, ApsRollingForecastFactoryCapacityDto.class))
+        .collect(Collectors.toList());
     ((ApsRollingForecastFactoryCapacityService) AopContext.currentProxy()).setName(dataList);
     return new ApsRollingForecastFactoryCapacityQueryListRes().setDataList(dataList);
   }
 
-  public @Override DynamicsPage<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> queryPageList(ApsRollingForecastFactoryCapacityExportQueryPageListReq req) {
+  public @Override DynamicsPage<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> queryPageList(
+      ApsRollingForecastFactoryCapacityExportQueryPageListReq req) {
 
     DynamicsPage<ApsRollingForecastFactoryCapacity> page = new DynamicsPage<>();
     page.setCurrent(req.getPageNum()).setSize(req.getPageSize());
@@ -78,15 +97,18 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
     List<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> records;
     if (Boolean.TRUE.equals(req.getQueryPage())) {
       IPage<ApsRollingForecastFactoryCapacity> list = this.page(page, q);
-      IPage<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> dataList = list.convert(t -> $.copy(t, ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class));
+      IPage<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> dataList = list.convert(
+          t -> $.copy(t, ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class));
       records = dataList.getRecords();
     } else {
-      records = $.copyList(this.list(q), ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class);
+      records = $.copyList(this.list(q),
+          ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class);
     }
 
     // 类型转换，  更换枚举 等操作
 
-    List<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> listInfoRes = $.copyList(records, ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class);
+    List<ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes> listInfoRes = $.copyList(
+        records, ApsRollingForecastFactoryCapacityExportQueryPageListInfoRes.class);
     ((ApsRollingForecastFactoryCapacityService) AopContext.currentProxy()).setName(listInfoRes);
 
     return DynamicsPage.init(page, listInfoRes);
@@ -106,11 +128,14 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
 
   @Override
   @Transactional
-  public ApsRollingForecastFactoryCapacityInsertRes save(ApsRollingForecastFactoryCapacityInsertReq req) {
+  public ApsRollingForecastFactoryCapacityInsertRes save(
+      ApsRollingForecastFactoryCapacityInsertReq req) {
 
 //    req.setCapacityList(req.getCapacityList().stream().sorted(Comparator.comparing(ApsRollingForecastFactoryCapacityDtoInfo::getBeginTime)).toList());
 
-    List<LocalDate> minMaxDate = DateUtils.getMinMaxDate(req.getCapacityList().stream().map(t -> List.of(t.getBeginTime(), t.getEndTime())).toList());
+    List<LocalDate> minMaxDate = DateUtils.getMinMaxDate(
+        req.getCapacityList().stream().map(t -> List.of(t.getBeginTime(), t.getEndTime()))
+            .toList());
 
     Map<String, ApsRollingForecastFactoryCapacity> capacityMap = Maps.newHashMap();
 
@@ -120,38 +145,46 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
 
     List<YearMonth> monthList = DateUtils.getMonthList(beginDate, endDate);
     $.requireNonNullCanIgnoreException(monthList, "时间范围获取失败,请检查时间范围");
-    List<ApsRollingForecastFactoryCapacity> factoryCapacityList = this.list(new LambdaQueryWrapper<ApsRollingForecastFactoryCapacity>()//
-        .eq(ApsRollingForecastFactoryCapacity::getFactoryId, req.getFactoryId())//
-        .and(r -> {
-          monthList.forEach((ym) -> {
-            r.or(rt -> {
-              rt.eq(ApsRollingForecastFactoryCapacity::getYear, ym.getYear()).eq(ApsRollingForecastFactoryCapacity::getMonth, ym.getMonth());
-            });
-          });
-        }));
+    List<ApsRollingForecastFactoryCapacity> factoryCapacityList = this.list(
+        new LambdaQueryWrapper<ApsRollingForecastFactoryCapacity>()//
+            .eq(ApsRollingForecastFactoryCapacity::getFactoryId, req.getFactoryId())//
+            .and(r -> {
+              monthList.forEach((ym) -> {
+                r.or(rt -> {
+                  rt.eq(ApsRollingForecastFactoryCapacity::getYear, ym.getYear())
+                      .eq(ApsRollingForecastFactoryCapacity::getMonth, ym.getMonth());
+                });
+              });
+            }));
     factoryCapacityList.forEach(t -> {
       capacityMap.put(t.getYear() + "-" + t.getMonth(), t);
     });
 
     req.getCapacityList().forEach(t -> {
-      List<LocalDate> localDateBetween = DateUtils.getLocalDateBetween(t.getBeginTime(), t.getEndTime());
+      List<LocalDate> localDateBetween = DateUtils.getLocalDateBetween(t.getBeginTime(),
+          t.getEndTime());
       localDateBetween.forEach(t1 -> {
         String key = t1.getYear() + "-" + t1.getMonthValue();
         ApsRollingForecastFactoryCapacity apsRollingForecastFactoryCapacity = capacityMap.get(key);
         if (Objects.isNull(apsRollingForecastFactoryCapacity)) {
           apsRollingForecastFactoryCapacity = new ApsRollingForecastFactoryCapacity();
-          apsRollingForecastFactoryCapacity.setFactoryId(req.getFactoryId()).setYear(t1.getYear()).setMonth(t1.getMonthValue());
+          apsRollingForecastFactoryCapacity.setFactoryId(req.getFactoryId()).setYear(t1.getYear())
+              .setMonth(t1.getMonthValue());
         }
-        ReflectUtil.setFieldValue(apsRollingForecastFactoryCapacity, "day" + ((t1.getDayOfMonth() < 10) ? "0" + t1.getDayOfMonth() : t1.getDayOfMonth()), t.getCapacity());
+        ReflectUtil.setFieldValue(apsRollingForecastFactoryCapacity,
+            "day" + ((t1.getDayOfMonth() < 10) ? "0" + t1.getDayOfMonth() : t1.getDayOfMonth()),
+            t.getCapacity());
         capacityMap.put(key, apsRollingForecastFactoryCapacity);
       });
     });
 
     Collection<ApsRollingForecastFactoryCapacity> values = capacityMap.values();
     if (CollUtil.isNotEmpty(values)) {
-      List<ApsRollingForecastFactoryCapacity> insertList = values.stream().filter(t -> Objects.isNull(t.getId())).toList();
+      List<ApsRollingForecastFactoryCapacity> insertList = values.stream()
+          .filter(t -> Objects.isNull(t.getId())).toList();
       this.saveBatch(insertList);
-      List<ApsRollingForecastFactoryCapacity> updateList = values.stream().filter(t -> Objects.nonNull(t.getId())).toList();
+      List<ApsRollingForecastFactoryCapacity> updateList = values.stream()
+          .filter(t -> Objects.nonNull(t.getId())).toList();
       this.updateBatchById(updateList);
     }
 
@@ -164,27 +197,33 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
     if (CollUtil.isEmpty(monthList)) {
       return null;
     }
-    List<ApsRollingForecastFactoryCapacity> factoryCapacityList = this.list(new LambdaQueryWrapper<ApsRollingForecastFactoryCapacity>()//
-        .eq(ApsRollingForecastFactoryCapacity::getFactoryId, factoryId).and(r -> {
-          monthList.forEach(t -> {
-            r.or(rt -> {
-              rt.eq(ApsRollingForecastFactoryCapacity::getYear, t.getYear()).eq(ApsRollingForecastFactoryCapacity::getMonth, t.getMonth());
-            });
-          });
-        }));
+    List<ApsRollingForecastFactoryCapacity> factoryCapacityList = this.list(
+        new LambdaQueryWrapper<ApsRollingForecastFactoryCapacity>()//
+            .eq(ApsRollingForecastFactoryCapacity::getFactoryId, factoryId).and(r -> {
+              monthList.forEach(t -> {
+                r.or(rt -> {
+                  rt.eq(ApsRollingForecastFactoryCapacity::getYear, t.getYear())
+                      .eq(ApsRollingForecastFactoryCapacity::getMonth, t.getMonth());
+                });
+              });
+            }));
     if (CollUtil.isEmpty(factoryCapacityList)) {
       return null;
     }
     List<LocalDate> localDateBetween = DateUtils.getLocalDateBetween(beginDate, endDate);
-    Map<String, ApsRollingForecastFactoryCapacity> capacityMap = factoryCapacityList.stream().collect(Collectors.toMap(t -> t.getYear() + "-" + t.getMonth(), Function.identity()));
+    Map<String, ApsRollingForecastFactoryCapacity> capacityMap = factoryCapacityList.stream()
+        .collect(Collectors.toMap(t -> t.getYear() + "-" + t.getMonth(), Function.identity()));
 
     List<FactoryCapacityDay> dayList = new ArrayList<>();
     localDateBetween.forEach(t -> {
-      ApsRollingForecastFactoryCapacity apsRollingForecastFactoryCapacity = capacityMap.get(t.getYear() + "-" + t.getMonthValue());
+      ApsRollingForecastFactoryCapacity apsRollingForecastFactoryCapacity = capacityMap.get(
+          t.getYear() + "-" + t.getMonthValue());
       if (Objects.nonNull(apsRollingForecastFactoryCapacity)) {
-        String fieldName = "day" + ((t.getDayOfMonth() < 10) ? "0" + t.getDayOfMonth() : t.getDayOfMonth());
+        String fieldName =
+            "day" + ((t.getDayOfMonth() < 10) ? "0" + t.getDayOfMonth() : t.getDayOfMonth());
         Field field = getField(apsRollingForecastFactoryCapacity, fieldName);
-        Integer capacity = (Integer) FieldUtils.getFieldValue(apsRollingForecastFactoryCapacity, field);
+        Integer capacity = FieldUtils.getFieldValue(apsRollingForecastFactoryCapacity,
+            field);
         capacity = Objects.isNull(capacity) ? 0 : capacity;
         dayList.add(new FactoryCapacityDay().setCapacity(capacity).setLocalDate(t));
       }
@@ -192,7 +231,8 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
     return dayList;
   }
 
-  private MPJLambdaWrapper<ApsRollingForecastFactoryCapacity> getWrapper(ApsRollingForecastFactoryCapacityDto obj) {
+  private MPJLambdaWrapper<ApsRollingForecastFactoryCapacity> getWrapper(
+      ApsRollingForecastFactoryCapacityDto obj) {
     MPJLambdaWrapper<ApsRollingForecastFactoryCapacity> q = new MPJLambdaWrapper<>();
 
     if (Objects.nonNull(obj)) {
@@ -202,7 +242,8 @@ public class ApsRollingForecastFactoryCapacityServiceImpl extends MPJBaseService
       LambdaQueryUtil.lambdaQueryWrapper(q, copy, ApsRollingForecastFactoryCapacity::getFactoryId);
     }
     q.orderByDesc(ApsRollingForecastFactoryCapacity::getId);
-    List<SFunction<ApsRollingForecastFactoryCapacity, ?>> columns = List.of(ApsRollingForecastFactoryCapacity::getYear, ApsRollingForecastFactoryCapacity::getMonth);
+    List<SFunction<ApsRollingForecastFactoryCapacity, ?>> columns = List.of(
+        ApsRollingForecastFactoryCapacity::getYear, ApsRollingForecastFactoryCapacity::getMonth);
     q.orderByDesc(columns);
     return q;
